implemented and tested working xdist lock

+++ pytestomatio/utils/worker_sync.py
@@ -13,14 +13,8 @@ class SyncLock:
         self.worker_id = str(uuid.uuid4())
         self.test_run_id: str or None = None
 
-    def lock(self) -> bool:
-        '''
-        :return: True if this worker is the first to lock the file, False otherwise
-        '''
-        time.sleep(random.randint(1, 1000) / 1000)
-        is_first = len([file for file in os.listdir() if file.startswith(prefix)]) == 0
+    def lock(self) -> None:
         open(prefix + self.worker_id, 'x').close()
-        return is_first
 
     def save_run_id(self, run_id: str) -> None:
         try:
@@ -35,33 +29,34 @@ class SyncLock:
     def get_run_id(self) -> str or None:
         if self.test_run_id is not None:
             return self.test_run_id
-
-        time.sleep(fixed_delay)
         try:
-            self.test_run_id = self._read_run_file()
+            with open(sync_run_file, 'r') as f:
+                self.test_run_id = f.read()
             return self.test_run_id
         except FileNotFoundError:
-            time.sleep(fixed_delay)
-            try:
-                self.test_run_id = self._read_run_file()
-                return self.test_run_id
-            except FileNotFoundError:
-                return None
+            return None
 
-    def _read_run_file(self) -> str:
-        with open(sync_run_file, 'r') as f:
-            return f.read()
-
-    def unlock(self) -> bool:
-        '''
-        :return: True if this worker is the last to unlock the file, False otherwise
-        '''
+    def unlock(self) -> None:
         os.remove(prefix + self.worker_id)
-        is_last = len([file for file in os.listdir() if file.startswith(prefix)]) == 0
-        return is_last
 
-    def clear_run_id(self) -> None:
+    def clear_run_id(self) -> bool:
+        file_count = len([file for file in os.listdir() if file.startswith(prefix)])
+        if file_count == 0 and not os.path.exists(sync_run_file):
+            return True
+        try:
+            with open(sync_run_file, 'w'):
+                for _ in range(10_000):
+                    file_count = len([file for file in os.listdir() if file.startswith(prefix)])
+                    if file_count == 0:
+                        break
+                    time.sleep(fixed_delay)
+        except PermissionError:
+            return False
+        except FileNotFoundError:
+            return True
         try:
             os.remove(sync_run_file)
-        except FileNotFoundError:
-            pass
+        except PermissionError:
+            return False
+        time.sleep(0.1)
+        return True
